Refactor: Organize screen files into subdirectories
Moved screen files into more specific subdirectories (event, organization, speaker, sponsor, agenda, login) to improve project structure and organization. Updated import paths accordingly.
