+++ lib/ui/screens/agenda/agenda_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:sec/ui/dialogs/delete_dialog.dart';
 
-import '../../core/models/models.dart';
-import '../../core/utils/date_utils.dart';
+import '../../../core/models/models.dart';
+import '../../../core/utils/date_utils.dart';
 
 class ExpansionTileState {
   final bool isExpanded;
