+++ lib/ui/screens/event/event_form/event_form_screen.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:sec/core/config/app_decorations.dart';
 import 'package:sec/core/config/app_fonts.dart';
 
-import '../../core/models/agenda.dart';
-import '../../core/utils/time_utils.dart';
-import '../widgets/widgets.dart';
+import '../../../../core/models/agenda.dart';
+import '../../../../core/utils/time_utils.dart';
+import '../../../widgets/widgets.dart';
 
 class EventFormData {
   final List<String> rooms;
