+++ lib/ui/screens/speaker/speaker/speakers_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:sec/ui/screens/speaker_form_screen.dart';
+import 'package:sec/ui/screens/speaker/speaker_form/speaker_form_screen.dart';
 
-import '../../core/models/models.dart';
-import '../../l10n/app_localizations.dart';
-import '../widgets/widgets.dart';
+import '../../../../core/models/models.dart';
+import '../../../../l10n/app_localizations.dart';
+import '../../../widgets/widgets.dart';
 
 /// Screen that displays a grid of speakers with their information and social links
 /// Fetches speaker data from the configured data source and displays it in cards
